requirements.txt
pandas==2.2.2
numpy==1.26.4
pyyaml==6.0
pytest==7.4.2
python-dateutil==2.8.2
